+++ src/steps/agents/index.test.ts
@@ -1,0 +1,24 @@
+import {
+  executeStepWithDependencies,
+  Recording,
+} from '@jupiterone/integration-sdk-testing';
+import { buildStepTestConfigForStep } from '../../../test/config';
+import { setupProjectRecording } from '../../../test/recording';
+import { Steps } from '../constants';
+
+let recording: Recording;
+
+afterEach(async () => {
+  await recording.stop();
+});
+
+test('#fetchAgents', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'fetch-agents',
+  });
+
+  const stepConfig = buildStepTestConfigForStep(Steps.AGENTS);
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
+});