+++ src/steps/clusters/converters.test.ts
@@ -1,0 +1,8 @@
+import { getMockCluster } from '../../../test/mocks';
+import { createClusterEntity } from './converter';
+
+describe('#createClusterEntity', () => {
+  test('should convert to entity', () => {
+    expect(createClusterEntity(getMockCluster())).toMatchSnapshot();
+  });
+});